refactor(grafana-ui): update imports to non-barrel files for story and test
