+++ packages/grafana-ui/src/components/Forms/FieldSet.story.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryFn } from '@storybook/react';
 
-import { Button } from '../Button';
+import { Button } from '../Button/Button';
 import { Input } from '../Input/Input';
 
 import { Field } from './Field';
