+++ packages/grafana-ui/src/components/InfoTooltip/InfoTooltip.tsx
@@ -1,5 +1,6 @@
 import { IconButton } from '../IconButton/IconButton';
-import { TooltipProps, PopoverContent } from '../Tooltip';
+import { TooltipProps } from '../Tooltip/Tooltip';
+import { PopoverContent } from '../Tooltip/types';
 
 interface InfoTooltipProps extends Omit<TooltipProps, 'children' | 'content'> {
   children: PopoverContent;
