+++ packages/grafana-ui/src/components/Input/Input.story.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { KeyValue } from '@grafana/data';
 
 import { getAvailableIcons } from '../../types/icon';
-import { Button } from '../Button';
+import { Button } from '../Button/Button';
 import { Field } from '../Forms/Field';
 
 import { Input } from './Input';
