+++ packages/grafana-ui/src/components/Modal/Modal.story.tsx
@@ -4,8 +4,8 @@ import { oneLineTrim } from 'common-tags';
 import { useState } from 'react';
 
 import { getAvailableIcons } from '../../types/icon';
-import { Button } from '../Button';
-import { TabContent } from '../Tabs';
+import { Button } from '../Button/Button';
+import { TabContent } from '../Tabs/TabContent';
 
 import { Modal } from './Modal';
 import mdx from './Modal.mdx';
