+++ packages/grafana-ui/src/components/PageLayout/PageToolbar.story.tsx
@@ -4,7 +4,7 @@ import { Meta } from '@storybook/react';
 import { StoryExample } from '../../utils/storybook/StoryExample';
 import { IconButton } from '../IconButton/IconButton';
 import { Stack } from '../Layout/Stack/Stack';
-import { ToolbarButton } from '../ToolbarButton';
+import { ToolbarButton } from '../ToolbarButton/ToolbarButton';
 
 import { PageToolbar } from './PageToolbar';
