+++ packages/grafana-ui/src/components/PanelChrome/PanelChrome.story.tsx
@@ -7,7 +7,7 @@ import { useInterval, useToggle } from 'react-use';
 import { LoadingState } from '@grafana/data';
 
 import { DashboardStoryCanvas } from '../../utils/storybook/DashboardStoryCanvas';
-import { Button } from '../Button';
+import { Button } from '../Button/Button';
 import { RadioButtonGroup } from '../Forms/RadioButtonGroup/RadioButtonGroup';
 import { Icon } from '../Icon/Icon';
 import { Stack } from '../Layout/Stack/Stack';
