+++ packages/grafana-ui/src/components/ThemeDemos/ThemeDemo.tsx
@@ -7,7 +7,7 @@ import tinycolor from 'tinycolor2';
 import { GrafanaTheme2, ThemeRichColor, ThemeVizHue } from '@grafana/data';
 
 import { useTheme2 } from '../../themes/ThemeContext';
-import { allButtonVariants, Button } from '../Button';
+import { allButtonVariants, Button } from '../Button/Button';
 import { Card } from '../Card/Card';
 import { CollapsableSection } from '../Collapse/CollapsableSection';
 import { Combobox } from '../Combobox/Combobox';
