+++ packages/grafana-ui/src/components/Toggletip/Toggletip.test.tsx
@@ -1,7 +1,7 @@
 ﻿import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { Button, LinkButton } from '../Button';
+import { Button, LinkButton } from '../Button/Button';
 
 import { Toggletip } from './Toggletip';
